+++ support-frontend/assets/components/thankYou/whatNext/activateSubscriptionReminder.tsx
