+++ support-frontend/assets/components/thankYou/whatNext/whatNext.tsx
